requirements.txt
python-dotenv==1.2.2
google-genai==1.67.0
streamlit==1.55.0
